==> tfm-disruptors/notebooks/etl/01_ETL_BBDD_EDlist.ipynb <==
# ETL BBDD EDlist
El objetivo de esta ETL es **extraer** las 3 listas que tiene la Endocrine Disruptor List https://edlists.org/ **transformar** cada BBDD para luego unir las 3 listas y posteriormente **cargar** usar este dataset transformado para otros fines que explicaremos más adelante. 

Para ello tenenmos que tener en cuenta que en la EDlist se explican los 3 listados: https://edlists.org/the-ed-lists

The three lists are:
* List I: Substances identified as endocrine disruptors at EU level
* List II: Substances under evaluation for endocrine disruption under an EU legislation
* List III: Substances considered, by the evaluating National Authority, to have endocrine disrupting properties

Tambien explica que con respecto a la lista III: "It should be noted that these compounds are not necessarily considered as suspected EDs at the EU level."

Con esta información, determinamos para nuestro análisis que: 
1. Lo que está en la lista I, solo debería estar en esta porque ya ha pasado de la lista II de evaluación o de la lista III en la que algún país miembro lo considera ED, pero el resto de la UE no. 
2. La lista II, puede contener ingredientes de la lista I que pese a estar en la lista I no se hayan eliminado de esta, o ingredientes de la lista III porque sigue en evaluación por parde de la UE, independientemente de que el país miembro que lo ha propuesto lo considere ED en su país. Si aparece en el archivo de esta lista las otras dos listas daremos la siguiente prioridad: lista I > lista II > lista III.
3. Lo que esté en la lista III, puede estar n la lista I pero aun no se haya eliminado de esta lista o tambien en la lista II porque esté en evaluación por la UE. Por tanto seguiremos la misma prioridad que en el apartado anterior.

## 1. Importamos librerias 

import pandas as pd
import numpy as np

## 2. Lectura de datos y estructuras de dataframes

# leemos los datasets de EDlist
edlist_1 = pd.read_excel("../../data/raw/list1.xlsx")
edlist_2 = pd.read_excel("../../data/raw/list2.xlsx")
edlist_3 = pd.read_excel("../../data/raw/list3.xlsx")

print (u"""
Nº de filas del dataset Edlist_1: {}
Nº de columnas del dataset Edlist_1: {}
Las variables de Edlist_1 son: {}

Nº de filas del dataset Edlist_2: {}
Nº de columnas del dataset Edlist_2: {}
Las variables de Edlist_2 son: {}

Nº de filas del dataset Edlist_3: {}
Nº de columnas del dataset Edlist_3: {}
Las variables de Edlist_3 son: {}

""".format(
    edlist_1.shape[0], edlist_1.shape[1], edlist_1.columns,
    edlist_2.shape[0], edlist_2.shape[1], edlist_2.columns,
    edlist_3.shape[0], edlist_3.shape[1], edlist_3.columns
    ))

edlist_1.head(2)

edlist_2.head(2)

edlist_3.head(2)

edlist_1.info()

edlist_2.info()

edlist_3.info()

## 3. Consulta de conflicto entre listas y mapeo de variables
Como se ha adelantado en el enunciado hay componentes que aparecen en otras listas, para ello consultamos la variable "Appears on lists"

# Consultamos los valores de "Appears on lists"
print("Edlist_1:")
print(edlist_1['Appears on lists'].value_counts(dropna= False))
print("\nEdlist_2:")
print(edlist_2['Appears on lists'].value_counts(dropna= False))
print("\nEdlist_3:")
print(edlist_3['Also appears on lists'].value_counts(dropna= False))

- En la lista I: hay 3 casos
- En la lista II: hay 22+3= 25 casos
- En la lista III: hay 22 casos

# Vemos los casos concreto de edlist_1
conflictos_1 = edlist_1[edlist_1['Appears on lists'].str.contains(",", na=False)]
print("Conflictos en Edlist_1:")
conflictos_1[['ID (for name)', 'Name and abbreviation', 'CAS no.', 'EC / List no.', 'Status', 'Status year', 'Appears on lists' ]]

Todos estos ingredientes han de ir en la Lista I porque ya se han demostrado que son ED.

Vemos que en realidad se trata solo de 1 ingrediente, pero hay varias líneas con información duplicada salvo la de Status y Status year, que para el objeto de estudio no es relevante.

Posiblemente pasará esto mismo con otros ingredientes, así que habrá que estudiar esto mismo más adelante.

# Vemos los casos concreto de edlist_2
conflictos_2 = edlist_2[edlist_2['Appears on lists'].str.contains(",", na=False)]
print("Conflictos en Edlist_2:")
conflictos_2[['ID (for name)', 'Name and abbreviation', 'CAS no.', 'EC / List no.', 'Status', 'Year', 'Appears on lists' ]]

De los 25 casos -> 3 deberían ser de la List I y el resto quedarse en la List II.

En realidad solo hay 6 ingredientes en conflicto, pero hay duplicidad.

# Vemos los casos concreto de edlist_3
conflictos_3 = edlist_3[edlist_3['Also appears on lists'].str.contains(",", na=False)]
print("Conflictos en Edlist_3:")
conflictos_3[['ID (for name)', 'Name and abbreviation', 'CAS no.', 'EC / List no.', 'Status', 'Year', 'Also appears on lists' ]]

De los 22 casos en conflicto -> todos deberían ser de la List II.

En realidad solo hay 5 ingredientes en conflicto, pero hay duplicidad.

**Renombramos categorías utilizando mapeo según el criterio establecido en el enunciado**

mapeo_categorias = {
    'List I, List II': 'List I',
    'List I, List II, List III': 'List I',
    'List II, List III': 'List II',
    'List I, List III': 'List I' # Esta combinación no existe actualmente pero se deja por si en algún futuro se da el caso.
}

edlist_1['Appears on lists'] = edlist_1['Appears on lists'].replace(mapeo_categorias)
edlist_2['Appears on lists'] = edlist_2['Appears on lists'].replace(mapeo_categorias)
edlist_3['Also appears on lists'] = edlist_3['Also appears on lists'].replace(mapeo_categorias)

# Comprobamos "Appears on lists"
print("Edlist_1:")
print(edlist_1['Appears on lists'].value_counts(dropna=False))
print("\nEdlist_2:")
print(edlist_2['Appears on lists'].value_counts(dropna=False))
print("\nEdlist_3:")
print(edlist_3['Also appears on lists'].value_counts(dropna=False))

Se han clasificado como queriamos.

## 4. Unión de tablas

Vamos a unir las tablas y nos vamos a quedar con todas las filas, porque no sabemos qué utilidad puedan tener más adelante.

En este punto podríamos eliminar duplicados, pero como seguramente hayan tambien duplicados no solo de la misma tabla sino entre distintas tablas, para no hacer el trabajo dos veces, dejamos ese análisis para después.

Antes de realizar la unicón vamos a:
- Unificar los nombres de columnas en todas las tablas para no tener problemas.
- Añadir una columna que nos diga la fuente de origen de la tabla

# renombramos columnas
edlist_1 = edlist_1.rename(columns={'Status year': 'Year'})
edlist_3 = edlist_3.rename(columns = {'Also appears on lists': 'Appears on lists'})

# añadimos columna
edlist_1['Fuente_original'] = 'Edlist_1'
edlist_2['Fuente_original'] = 'Edlist_2'
edlist_3['Fuente_original'] = 'Edlist_3'

# Comprobamos una de ellas
edlist_1.info()

# Ordenamos las 3 listas por ID (for name)
edlist_1 = edlist_1.sort_values( 'ID (for name)')
edlist_2 = edlist_2.sort_values( 'ID (for name)')
edlist_3 = edlist_3.sort_values( 'ID (for name)')

# Concatenamos
edlist_completa = pd.concat([edlist_1, edlist_2, edlist_3], ignore_index=True)

edlist_completa.info()

# Vemos que el índice se ha reseteado correctamente y hay continuidad del mismo de la List I a la List II
edlist_completa.iloc[225:227]

# Entre la List II y la III ocurre lo mismo
edlist_completa.iloc[329:331]

## 5. Tratamiento de Duplicados

Recordemos que solo nos es necesario tener un ingrediente, aunque aparezca en dos listas.

# Comprobamos nulos
edlist_completa.isna().sum()

# Para hacer la comparativa de duplicados vamos a elegir la variable "Name and abbreviation" pues es única para cada ingrediente y no tiene nulos.
duplicados = edlist_completa[edlist_completa.duplicated(subset='Name and abbreviation', keep=False)].sort_values(by=['ID (for name)']) 
duplicados.iloc[84:92]

Hemos revisado la lista de duplicados, hay más de 250 registros duplicados. 

Mostramos el ejemplo para el ingrediente con ID(for name) : 249 

Este tiene varias duplicidades:
1. Duplicidad inter lista: vemos que en la Edlist_2 viene el elemento 4 veces, como hemos dicho eso pasa porque hay un hitórico de Status.
2. Duplicidad entre listas: vemos que en la Edlist_3 aparecía 4 veces y ponía que aparecía en la lista II a la que le dabamos prioridad, porque sigue en estudio por la UE.

En este caso, nos quedaremos con el Status year más reciente, pero que nos de más información. 

# Consultamos duplicados por las variables más relevantes
duplicados = edlist_completa[edlist_completa.duplicated(subset=['ID (for name)','Name and abbreviation', 'CAS no.', 'EC / List no.'], keep='first')].sort_values(by=['ID (for name)']) 
duplicados.count()

151 sería el número de registros a eliminar

# Antes comprobmaso si el año es numérico
edlist_completa['Year'].info ()

# Tambien comprobamos que hay un solo valor para cada año
edlist_completa['Year'].value_counts()

# Renombramos los años quendono con el año más reciente
año = {'2022, 2023': '2023',
        '2020, 2021': '2021'}

edlist_completa['Year'] = edlist_completa['Year'].replace(año)


# Convertimos la variable en numérica
edlist_completa['Year'] = pd.to_numeric(edlist_completa['Year'])

print(edlist_completa['Year'].value_counts()) 

Avanzando en el código nos dimos cuenta que si solo nos quedamos con el año más reciente, la variable Regulatory Field se quedan sin información.

Por ello vamos a añadir una variable boleanas para ver si hay nulos.

Con ello, al eliminar duplicados tendremos en cuenta estas variable que son relevantes y de las que necesitaremos información más adelante y nuestro dataframe no se quedará con nulos innecesariamente.

# Eliminamos los duplicados por las variables más relantes y ordenando por las otras que pierdan información
edlist_final = edlist_completa.drop_duplicates(
    subset=['ID (for name)','Name and abbreviation', 'CAS no.', 'EC / List no.'], keep='first').sort_values(by = ['Year', 'Regulatory Field', 'Environmental effects','Health effects'],
    ascending=[ True, False, False, False ])
# Ordenamos de nuevo por "ID (for name)"
edlist_final= edlist_final.sort_values(by='ID (for name)')


# Comprobamos que efectivamente, el caso 249 se ha quedado con el que más información da y el más reciente
edlist_final.iloc[54:57] 

**Eliminamos varibles creadas para la ordenación**

edlist_final.info()

## 6. Tratamiento de Nulos de determinadas variables.

La variables **Environmental effects** y **Health effects** tiene como valor NAN los ceros, por tanto vamos a darles valor : 0-1
No tratamos el resto de nulos, porque eso ya es parte del EDA y previo a esto hay que tratarlo con más dataframes.

# Mapear 'Yes' a 1, y el resto a 0 (incluye NaN)
edlist_final['Health effects'] = (edlist_final['Health effects'] == 'Yes').astype(int)
edlist_final['Environmental effects'] = (edlist_final['Environmental effects'] == 'Yes').astype(int)


# Como hemos mencionado antes esta es la combinación en la que el dataframe se queda con los menores nulos posibles.
edlist_final.isnull().sum()

## 7. Variables "CAS no." y "EC / List no."

Estas son variables comunes con el dataframe ECHA y son las claves por la que vamos a realizar la unión de tablas con otra fuente de datos.

Se ha detectado que hay registros con varios "CAS no." y "EC / List no." vamos a investigar cuáles son para darle tratamiento antes de unir.

# Filtramos filas con más de un CAS no. 
edlist_final[edlist_final['CAS no.'].str.contains(r'[/;,]', regex= True, na=False)]

Vemos que en CAS no. hay 4 registros que tienen un varios CAS no.

Ahora veremos en EC / List no. coinciden los 4 y se añade uno más.

# Filtramos filas con más de un EC / List no.
edlist_final[edlist_final['EC / List no.'].str.contains(r'[/;,]', regex= True, na=False)]

# Separamos CAS múltiples por coma en una lista
edlist_final['CAS_list'] = edlist_final['CAS no.'].str.split(r'[,;/]')
edlist_final['EC_list'] = edlist_final['EC / List no.'].str.split(r'[,;/]')
edlist_final[['ID (for name)','CAS_list','EC_list' ]].iloc[27:33] # localizamos algunos resultados

Descomponemos cada elemento de la lista de CAS_list y de EC_list. Al hacerlo se van a crear combinaciones entre ambas porque se va a descomponer hasta que cada una tenga solo un elemento.

De est manera si CAS_list tiene 4 elementos se tendrían que crear 4 filas para cada CAST_list pero si además EC_list tiene otros 4 elementos, no serían 8, sino 4^4 =16 porque se combinan los índices de las listas, como es en el caso del ID(for name ): 179

# Descomponemos fila por cada CAS individual
edlist_final = edlist_final.explode('CAS_list')
# Descomponemos fila por cada EC individual
edlist_final = edlist_final.explode('EC_list')

Comprobamos el resultado y vemos que se han realizado correctamente todas las combinaciones

# Vemos resultados de 'EC / List no.' porque incluye tambien los resultados de 'CAS no.'
consulta = edlist_final[edlist_final['EC / List no.'].str.contains(r'[/;,]', regex= True,na=False)]
consulta[['Name and abbreviation', 'CAS no.', 'EC / List no.', 'CAS_list', 'EC_list']]. head(16)

Dejamos las variables nuevas así aunque sean el mismo ingrediente duplicado, porque luego queremos unir las tablas y entre más información tenga la variable mejor será la unión entre tablas.

# Eliminamos 'CAS no.' y 'EC / List no.' porque ya no nos sirven.
edlist_final.drop(columns='CAS no.', inplace=True)
edlist_final.drop(columns='EC / List no.', inplace=True)

#renombrar variables con el nombre antiguo
edlist_final = edlist_final.rename(columns={'CAS_list': 'CAS Number', 'EC_list':'EC Number'})

Comprobamos que no hay valores extraños en las variables "EC no." y "CAS no." antes de exportar

edlist_final[edlist_final['CAS Number'] == '-']

edlist_final['CAS Number'] = edlist_final['CAS Number'].replace(['-', '', ' '], np.nan)
edlist_final['EC Number'] = edlist_final['EC Number'].replace(['-', '', ' '], np.nan)

Para no crear conflictos en la unión, tenemos que comprobar que las 2 varibles clave: "CAS no." y "EC no." no tengan nulos.

# Comprobamos la condición de que "CAS no." y "EC no." sean nulos.

edlist_final[(edlist_final['CAS Number'].isna() & edlist_final['EC Number'].isna())]

# Filtramos los registros que cumplen la condición que "CAS no." y "EC no." sean NO nulos.

edlist_final = edlist_final[~(edlist_final['CAS Number'].isna() & edlist_final['EC Number'].isna())]

edlist_final.info()

## 8. Exportamos base de datos limpia
Posteriormente vamos a unir esta tabla con la tabla ECHA y CosIng.

# Reiniciamos ínfice
edlist_final.reset_index(drop=True, inplace=True)
edlist_clean= edlist_final

# CSV
# edlist_clean.to_csv(""../../data/processed/notebooks/edlist_clean.csv", index=False)

# Parquet
edlist_clean.to_parquet("../../data/processed/notebooks/edlist_clean.parquet", index=False)
