==> tfm-disruptors/notebooks/etl/02_ETL_BBDD_ECHA.ipynb <==
# ETL BBDD ECHA
El objetivo de esta ETL es **extraer** el listado de Disruptores Endocrinos de la ECHA (European Chemicals Agency) https://echa.europa.eu/es/ed-assessment **transformar** la BBDD para luego **cargarla** junto con la Edlist y unir ambos dataframes.

El motivo de evaluar esta BBDD y no solo quedarnos con la EDlist, en la página de la Edlist se informa:
https://edlists.org/ed-relevant-legislations

" Más de 40 legislaciones diferentes intervienen en la regulación de sustancias químicas en la UE. Aquí se ofrece una breve descripción de las jurisdicciones más relevantes para la regulación de sustancias químicas disruptoras endocrinas, así como de los procedimientos correspondientes.
- Organismo responsable : ECHA (Agencia Europea de Sustancias y Mezclas Químicas)
- Jurisdicción : REACH regula los productos químicos industriales producidos o importados en la UE.

Procedimiento : Si un Estado miembro o la ECHA tienen una preocupación debido a las propiedades de alteración endocrina de un compuesto, pueden proponerlo como SVHC (Sustancia Extremadamente Preocupante) de acuerdo con el artículo 57(f) del reglamento REACH. Tras una ronda de comentarios públicos, la propuesta se remite al Comité de los Estados Miembros (MSC) para la búsqueda de un acuerdo. Si el MSC alcanza una decisión unánime, el compuesto se incluye en la Lista de Sustancias Candidatas de Muy Alta Preocupación para su Autorización debido a sus propiedades de alteración endocrina. Si los países del MSC no están de acuerdo, la propuesta se remite a la Comisión Europea para una decisión final.

(...)

La ECHA evalúa periódicamente las sustancias de la Lista de Candidatos para decidir cuáles deben incluirse en la Lista de Autorización con carácter prioritario. "

Por ese motivo, consideramos que para cubrir un mayor campo, hay que tener en cuenta ambos listados y en el peor de los caso eliminar duplicados.

La ECHA explica en su web (enlace anterior) en el apartado "Explanatory note":

"La inclusión en la lista significa que se está desarrollando o se ha completado desde febrero de 2013 una evaluación informal de riesgos para las propiedades disruptoras endocrinas.

Para cada sustancia, la tabla muestra el Estado miembro evaluador (presentador), el resultado de la evaluación y la fecha de la última actualización de la entrada de la lista."

Por ese motivo, tendrémos en cuenta las variables "Status" y "Outcome" en donde en combinación de ambas variables se manifiesta si es disruptor confirmado o en evaluación

## 1. Importamos librerias

import pandas as pd
import numpy as np

## 2. Lectura de datos e información de la lista ECHA

# leemos los datos de la lista de ECHA
echa= pd.read_excel("../../data/raw/endocrine-disruptor-assessment-export.xlsx", skiprows=3)


print("""
Nº de filas del dataset ECHA: {}
Nº de columnas del dataset ECHA: {}
Las variables de ECHA son: {}
""".format(echa.shape[0], echa.shape[1], echa.columns.tolist()))

echa.info()

Solo hay 1 nulo en **Authority** y casi todos los de la variable **DISLIST_ED_description** son nulos.

El resto de variables parecen estar completas.

**Procedemos a su eliminación "DISLIST_ED_description"**
Porque no contiene apenas información y además aunque en el archivo aparezca en la web no aparece.

echa.drop(columns='DISLIST_ED_description', inplace=True)

# Renombramos variable para que coincida con la Edllist
echa.rename(columns={'CAS no': 'CAS Number', 'EC / List no':'EC Number'}, inplace=True)

## 3. Duplicados

# Para hacer la comparativa de duplicados vamos a elegir la variable "Name and abbreviation" pues es única para cada ingrediente y no tiene nulos.
duplicados = echa[echa.duplicated(subset=['Substance name', 'CAS Number', 'EC Number' ], keep=False)]
duplicados

# Eliminamos los duplicados por las variables más relantes y ordenando por las otras que pierdan información
echa = echa.drop_duplicates(subset=['Substance name', 'CAS Number', 'EC Number' ], keep='first').sort_values(by = 'Authority')

echa.info()

## 4. Comprobación de "EC Number" y "CAS Number" - Mismo procedimiento que para EDlist

Las dos variables que coinciden en ambos dataframes son "EC Number" y "CAS Number", por tanto vamos a comprobar el estado de esas variables en el dataframe ECHA:

Comprobamos si hay más de un "EC Number" y "CAS Number" en el mismo registro, como pasaba con la Edlist

# Filtrar filas con más de un CAS no. de ECHA
echa[echa['CAS Number'].str.contains(r'[/;,]', regex= True,na=False)]

# En este dataframe no hay problemas

# Filtrar filas con más de un CAS no. de ECHA
echa[echa['EC Number'].str.contains((r'[,;/]'), na=False)]

# En este dataframe no hay problemas

echa[echa['CAS Number'] == '-']

Vemos que ambas variables "EC Number" y "CAS Number" contienen valores: "-" por lo que los sustituiremos por NAN

echa['CAS Number'] = echa['CAS Number'].replace(['-', '', ' '], np.nan)
echa['EC Number'] = echa['EC Number'].replace(['-', '', ' '], np.nan)

echa.info()

Vemos que hemos pasado de no tener nulos en las variables "EC no." y "CAS no." ahora tenemos algunos.
Tambien vemos que en la consulta anterior hay dos registros que tienen en nulo en las 2 varibles que por las que vamos a hacer la unión: "CAS no." y "EC no."
Hay que eliminar para no crear conflictos en la unión.

# Filtramos los registros que umplen la condición que "CAS no." y "EC no." sean NO nulos.

echa_clean = echa[~(echa['CAS Number'].isna() & echa['EC Number'].isna())]

echa_clean.info()

## 5. Creación variable fuente_original

Vamos a crear esta variable con la información según el disruptor está en evaluación o está confirmado

echa_clean['Status'].value_counts()

echa_clean['Outcome'].value_counts()

La información de la variable Status se amplia en la variable Outcome, que es en donde se dice cómo se ha concluido la investigación.

Por ello, vamos consultar qué valores toma **Outcome** cuando la variable **Status** toma valor concluide y cuando no

Status_concluded = echa_clean[echa_clean['Status']== 'Concluded']
Status_concluded['Outcome'].value_counts()

Status_no_concluded = echa_clean[echa_clean['Status']!= 'Concluded']
Status_no_concluded['Outcome'].value_counts()

Con esta información vamos a mapear outcome para darle valores: 
* ED_Env
* ED_Humans
* ED_Humans&Env
* ED_pendiente

Tendremos en cuenta como pendiente inconclusive e postponed, porque si bien es cierto que están dentro del grupo de **concluded**, no se ha descartado ni confirmado como disruptor.

Pero antes vamos a elimar los que no son ED porque al estar confirmados que no lo son no nos interesa tenerlos

echa_clean= echa_clean[~echa_clean['Outcome'].isin(['not ED', 'not ED#inconclusive'])]
echa_clean['Outcome'].value_counts()

# Creamos un diccionario de mapeo
mapeo = {
    "ED ENV": "ECHA - ED_Env",
    "ED ENV#ED HH": "ECHA - ED_Humans&Env",
    "ED HH": "ECHA - ED_Humans",
    "ED ENV 1": "ECHA - ED_Env",
    "Under development (SEV)" : "ECHA - ED_pendiente",
    "Under development (BPR)" : "ECHA - ED_pendiente",
    "inconclusive" : "ECHA - ED_pendiente",
    "postponed" : "ECHA - ED_pendiente",
    "Under development (other)" : "ECHA - ED_pendiente",
    "ED ENV#inconclusive" : "ECHA - ED_pendiente"
}

# Creamos la nueva columna
echa_clean['Fuente_original'] = echa_clean['Outcome'].map(mapeo)


echa_clean.info()

echa_clean['Fuente_original'].value_counts()

## 6. Exportamos base de datos limpia
Posteriormente vamos a unir esta tabla con la tabla EDlist y CosIng.

# Reiniciamos ínfice
echa_clean.reset_index(drop=True, inplace=True)

# CSV
# echa_clean.to_csv("../../data/processed/notebooks/echa_clean.csv", index=False)

# Parquet
echa_clean.to_parquet("../../data/processed/notebooks/echa_clean.parquet", index=False)
