==> tfm-disruptors/notebooks/etl/03_ETL_BBDD_COSING.ipynb <==
# ETL BBDD COSING
El objetivo de esta ETL es **extraer** los 4 anexos que tiene la Comisión Europea  https://ec.europa.eu/growth/tools-databases/cosing/reference/annexes **transformar** cada BBDD para luego unir los 4 anexos y posteriormente **cargar** usar este dataset transformado para otros fines que explicaremos más adelante. 

Para ello tenemos que tener en cuenta que en el reglamento se explican los 6 anexos relativos a este texto legislativo : https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX%3A32009R1223&qid=1611669486511

- ANEXO II: LISTA DE SUSTANCIAS PROHIBIDAS EN PRODUCTOS COSMÉTICOS
- ANEXO III: LISTA DE SUSTANCIAS QUE LOS PRODUCTOS COSMÉTICOS NO DEBEN CONTENER SALVO LAS RESTRICCIONES ESTABLECIDAS
- ANEXO IV: LISTA DE COLORANTES PERMITIDOS EN PRODUCTOS COSMÉTICOS
- ANEXO V: LISTA DE CONSERVANTES PERMITIDOS EN PRODUCTOS COSMÉTICOS
- ANEXO VI: LISTA DE FILTROS UV PERMITIDOS EN PRODUCTOS COSMÉTICOS


Con esta información, determinamos para nuestro análisis que: 

La variable que queremos extraer de estos anexos es : ***Name of Common Ingredients Glossary*** 

Según el Reglamento (CE) n.º 1223/2009 del Parlamento Europeo y del Consejo, de 30 de noviembre de 2009, sobre los productos cosméticos
https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX:02009R1223-20240424
en el artículo 33. Glossary of common ingredient names:  establece que la Comisión actualizará un glosario de nombres comunes de ingredientes, en base a la Nomenclatura Internacional de Ingredientes Cosméticos INCI. El nombre común del ingrediente se aplicará a efectos del etiquetado de los productos cosméticos comercializados. 

**IMPORTANTE:** En adelante cuando nos refiramos al nombre INCI nos referimos al nombre de la etiqueta de los productos, por lo que acabamos de explicar

## 1. Importamos librerias 

#importación de librerias
import pandas as pd
import numpy as np

# Función para cargar un archivo y sobrescribir el header

# 1. Definimos los headers.
# El anexo II cuenta con varias columnas menos.
# El anexo IV cuenta con una columna más.

header_2 = ['Reference Number', 'Chemical name','CAS Number', 'EC Number', 'Regulation', 'Other Directives/Regulations', 
'SCCS opinions', 'Chemical/IUPAC Name', 'Identified INGREDIENTS or substances e.g.', 'CMR', 'Update date']

header = ['Reference Number', 'Chemical name', 'Name of Common Ingredients Glossary' ,
'CAS Number', 'EC Number', 'Product Type, body parts','Maximum concentration in ready for use preparation' , 'Other Restrictions','Wording of conditions of use and warnings',
'Regulation', 'Other Directives/Regulations', 'SCCS opinions', 'Chemical/IUPAC Name', 'Identified INGREDIENTS or substances e.g.', 'CMR', 'Update date']

header_4 = ['Reference Number', 'Chemical name', 'Name of Common Ingredients Glossary' ,
'CAS Number', 'EC Number','Color', 'Product Type, body parts','Maximum concentration in ready for use preparation' , 'Other Restrictions','Wording of conditions of use and warnings',
'Regulation', 'Other Directives/Regulations', 'SCCS opinions', 'Chemical/IUPAC Name', 'Identified INGREDIENTS or substances e.g.', 'CMR', 'Update date']

# 2. Definimos función
def cargar_anexo(ruta, columnas):
    df = pd.read_excel(ruta, skiprows=6, header=[0,1])
    df.columns = columnas  # renombra las columnas con las listas creadas
    return df


## 2. Lectura de datos y estructuras de dataframes

Hemos visto que los archivos tienen el encabezado en celdas combinadas.

# leemos datasets del Anexo 3 de CosIng
anexo_3 = pd.read_excel("../../data/raw/COSING_Annex_III_v2.xlsx", skiprows=6, header = [0,1])
anexo_3.head(1)

Con esta información decidimos hacer directamente la lectura con una función (definida previamente) que reemplace directamente los nombres de las variables.

# leemos los datasets de CosIng

anexo_2 = cargar_anexo("../../data/raw/COSING_Annex_II_v2.xlsx", header_2)
anexo_3 = cargar_anexo("../../data/raw/COSING_Annex_III_v2.xlsx", header)
anexo_4 = cargar_anexo("../../data/raw/COSING_Annex_IV_v2.xlsx", header_4)
anexo_5 = cargar_anexo("../../data/raw/COSING_Annex_V_v2.xlsx", header)
anexo_6 = cargar_anexo("../../data/raw/COSING_Annex_VI_v2.xlsx", header)

Consultamos datos y su estructura

anexo_3.head(1)

anexo_2.info()

anexo_3.info()

anexo_4.info()

anexo_5.info()

anexo_6.info()

## 3. Tratamiento de variables antes de unión

Antes de realizar la unión de los anexos vamos añadir una columnas que nos diga a qué anexo pertenece.

# añadimos columna
anexo_2['Anexo_cosIng'] = 'Anexo_2'
anexo_3['Anexo_cosIng'] = 'Anexo_3'
anexo_4['Anexo_cosIng'] = 'Anexo_4'
anexo_5['Anexo_cosIng'] = 'Anexo_5'
anexo_6['Anexo_cosIng'] = 'Anexo_6'

# Concatenamos
cosing = pd.concat([anexo_2,anexo_3, anexo_4, anexo_5, anexo_6], ignore_index=True)

Consultamos estructura

cosing.info()

# Vemos que el índice se ha reseteado correctamente y hay continuidad del mismo del Anexo II al Anexo III
# Se aprecia tambien en la variable "Reference Number"
cosing.iloc[1743:1745]

## 4. Tratamiento de Duplicados

# Comprobamos nulos
cosing.isna().sum()

Antes de eliminar los duplicados, vamos a ordenar por variables relevantes que se pueden quedar sin información después de la eliminación.
Con ello, al eliminar duplicados tendremos en cuenta estas variable que son relevantes y de las que necesitaremos información más adelante y nuestro dataframe no se quedará con nulos innecesariamente.

# Para hacer la comparativa de duplicados vamos a elegir la variable "Chemical name" ya que es una variable importante y no tiene nulos.
duplicados = cosing[cosing.duplicated(subset=['Chemical name'], keep=False)].sort_values(by= ['Chemical name', 'Identified INGREDIENTS or substances e.g.'])
duplicados.head(8)

Vemos que hay ingredientes que han sido eliminados y no tienen información en las variables CAS Number y EC Number

# Eliminamos registros 'Moved or deleted'
cosing = cosing[~cosing['Chemical name'].str.contains('Moved or deleted', na=False)]

duplicados = cosing[cosing.duplicated(subset=['Chemical name'], keep=False)].sort_values(by= ['Chemical name', 'Identified INGREDIENTS or substances e.g.'])
duplicados.head(20)

Vemos que para esos dupicados las variables CAS Number y EC Number están tambien duplicadas, así pues eliminamos solo por variable 'Chemical name'

# Eliminamos los duplicados por 'Chemical name' y manteniendo el primero
cosing_final = cosing.drop_duplicates(subset='Chemical name' , keep='first')

Eliminamos las filas que continen en la variable "Chemical name" la categoría : Moved or deleted

# Consulta
cosing_final.isnull().sum()

## 5. Variables "'Name of Common Ingredients Glossary'"

Vamos a comprobar la varible 'Name of Common Ingredients Glossary', porque es la variable objetivo, pues es de la que obtendremos el nombre INCI.

Vamos a ver si esta variable tiene más de un nombre en un registro separado por comas, barras o punto y coma.

Los guiones y las comas no podemos tenerlos en cuenta porque claramente forman parte del nombre, y aunque parece más un nombre químico se encuentra en esta variable, por ejemplo: 

2-METHOXYMETHYL-P-PHENYLENEDIAMINE

1-Hydroxy-4-methyl-6-(2,4,4-trimethylpentyl)-2 pyridon

cosing_final[cosing_final['Name of Common Ingredients Glossary'].str.contains(r'[/;,]', regex= True, na=False)].count()

Sabemos que hay hasta 95 registros con estas condiciones, pero no vamos a tratarlos hasta después de la unión de las tablas porque sería agrandar innecesariamente esta tabla, cuando a lo mejor la mayoría de los registros no tienen match con la otra tabla. 

Una vez se haga la unión se revisará esta variable.

## 6. Variables "CAS Number" y "EC Number"

Estas son variables comunes con los dataframes Edlists y ECHA y son las claves por la que vamos a realizar la unión de tablas.

En los otros dataframes se detectaron registros con varios "CAS Number" y "EC Number" vamos a investigar si aquí ocurre lo mismo.

A diferencia de lo que pasaba con lo otros dataframes en este hay CAS Number separados por los simbolos: **','** **';'**  **'/'**

# Filtramos filas con más de un CAS Number
cosing_final[cosing_final['CAS Number'].str.contains(r'[,;/]', regex= True, na=False)].count()

Hay 212 registros que tienen un varios CAS Number

Ahora veremos que con EC Number pasa lo mismo

# Filtramos filas con más de un EC Number
cosing_final[cosing_final['EC Number'].str.contains(r'[,;/]', regex= True, na=False)].count()

# Separamos CAS y EC múltiples por coma en una lista
cosing_final['CAS Number_2'] = cosing_final['CAS Number'].str.split(r'[,;/]')
cosing_final['EC Number_2'] = cosing_final['EC Number'].str.split(r'[,;/]')

# comprobamos
cosing_final[cosing_final['CAS Number_2'].str.contains(r'[,;/]', regex= True, na=False)].count()

Descomponemos cada elemento de la lista de CAS Number_2 y de EC Number_2 Al hacerlo se van a crear combinaciones entre ambas porque se va a descomponer hasta que cada una tenga solo un elemento.

De esta manera si CAS Number_2 tiene 4 elementos se tendrían que crear 4 filas para cada CAS Number_2 pero si además EC Number_2 tiene otros 4 elementos, no serían 8, sino 4^4 =16 porque se combinan los índices de las listas.

# Descomponemos fila por cada CAS individual
cosing_final = cosing_final.explode('CAS Number_2')
# Descomponemos fila por cada EC individual
cosing_final = cosing_final.explode('EC Number_2')

#Comprobamos que 'CAS Number_2' y 'EC Number_2' tengan solo un número
cosing_final[['CAS Number','EC Number', 'CAS Number_2','EC Number_2' ]].iloc[15:30]

Eliminamos las variables anteriores que contienen el más de un número en estas variables

# Eliminamos 'CAS Number' y 'EC Number' porque ya no nos sirven.
cosing_final.drop(columns='CAS Number', inplace=True)
cosing_final.drop(columns='EC Number', inplace=True)

#renombrar variables con el nombre antiguo
cosing_final = cosing_final.rename(columns={'CAS Number_2': 'CAS Number', 'EC Number_2':'EC Number'})

Comprobamos que no hay valores extraños en las variables "EC no." y "CAS no." antes de exportar

cosing_final[cosing_final['CAS Number'] == '-'].count()

cosing_final[cosing_final['EC Number'] == '-'].count()

cosing_final['CAS Number'] = cosing_final['CAS Number'].replace(['-', '', ' '], np.nan)
cosing_final['EC Number'] = cosing_final['EC Number'].replace(['-', '', ' '], np.nan)

cosing_final.info()

Como lo único que queremos de esta tabla es localizar el nombre INCI nos quedaremos con las variables mínimas que nos ayuden a identificar el ingrediente y por ende no tenga tantos valores nulos

cosing_final= cosing_final[['Chemical name', 'Chemical/IUPAC Name', 'Identified INGREDIENTS or substances e.g.', 'CAS Number', 'EC Number','Name of Common Ingredients Glossary', 'Product Type, body parts', 'Anexo_cosIng' ]]

# Comprobamos la condición de que "CAS Number" y "EC Number" sean nulos.

cosing_final[(cosing_final['CAS Number'].isna() & cosing_final['EC Number'].isna())].count()

# Filtramos los registros que cumplen la condición que "CAS Number" y "EC Number" sean NO nulos. 
cosing_clean = cosing_final[~(cosing_final['CAS Number'].isna() & cosing_final['EC Number'].isna())]

cosing_clean.info()

cosing_clean['Chemical name'].nunique()

cosing_clean['CAS Number'].nunique()

cosing_clean['EC Number'].nunique()

# consultamos duplicados
duplicados = cosing_clean[
    cosing_clean.duplicated(
        subset=['Chemical name','EC Number','CAS Number', 'Identified INGREDIENTS or substances e.g.'], keep='first')].sort_values(
            by= ['Chemical name', 'Identified INGREDIENTS or substances e.g.','CAS Number','EC Number','Chemical/IUPAC Name'])
duplicados.head(10)

duplicados.count()

# Eliminamos los duplicados por las variables  de la consutla
cosing_clean = cosing_clean.drop_duplicates(subset=['Chemical name','EC Number','CAS Number', 'Identified INGREDIENTS or substances e.g.'] , keep='first').sort_values(
            by= ['Chemical name', 'Identified INGREDIENTS or substances e.g.','CAS Number','EC Number','Name of Common Ingredients Glossary','Chemical/IUPAC Name', 'Product Type, body parts'])
cosing_clean.info()

## 8. Exportamos base de datos limpia
Posteriormente vamos a unir esta tabla con la tabla ECHA y CosIng.

# Reiniciamos ínfice
cosing_clean.reset_index(drop=True, inplace=True)

Vemos que el indice no corresponde con los registros

# Parquet
cosing_clean.to_parquet("../../data/processed/notebooks/cosing_clean.parquet", index=False)

# Excel
# cosing_clean.to_excel("../../data/processed/notebooks/cosing_clean.xlsx", index=False)
