==> tfm-disruptors/notebooks/etl/04_ETL_BBDD_PESTICIDAS.ipynb <==
# ETL BBDD PETICIDAS_UE
El objetivo de esta ETL es **extraer** de la api que dispone la Unión Europea  https://data.europa.eu/data/datasets/pesticides?locale=en para **transformar** la BBDD para luego **cargar** y usar este dataset transformado para la unión de tablas

## 1. Importamos librerias 

import requests
import pandas as pd
import ast

## 2. Lectura de datos y estructuras de dataframes

# Extrción de datos de la API de la Unión Europea sobre pesticidas
headers = {
    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)',
    'Accept': 'application/json',
    'Referer': 'https://api.datalake.sante.service.ec.europa.eu'
}

url = "https://api.datalake.sante.service.ec.europa.eu/sante/pesticides/active_substances?format=json&api-version=v2.0"

response = requests.get(url, headers=headers)

if response.status_code == 200:
    data = response.json()
    pesticidas = pd.json_normalize(data['value'])
else:
    print("Error:", response.status_code, response.text[:200])


pesticidas.info()

## 3. ETL

# Consultamos las columnas que nos interesan
pesticidas =pesticidas [['substance_name', 'as_cas_number']]
pesticidas.head()

Consultamos los valores únicos de ambas variables

pesticidas['as_cas_number'].nunique(dropna=True)

pesticidas[pesticidas['as_cas_number'].str.contains(r'[,;/]', regex= True, na=False)].count()

# Nos quedamos con las variables que nos interesan
pesticidas = pesticidas[['substance_name', 'as_cas_number']]

# Renombramos las columnas
pesticidas.rename(columns={'substance_name': 'Name_Pesticida', 'as_cas_number': 'CAS Number'}, inplace=True)

# Exportamos el DataFrame 
pesticidas.to_parquet("../../data/processed/notebooks/pesticidas.parquet", index=False)
