==> tfm-disruptors/notebooks/etl/05_Merge_edlist_echa_pesticidas_cosing.ipynb <==
# UNIÓN DE TABLAS EDLIST - ECHA - PESTICIDAS - COSING
El objetivo de este archivo es la unión de tablas y la creación de una tabla única con el nombre que aparece en las etiquetas

Las tablas que se van a unir son:

- **edlist_clean** -> contiene disrupores endocrinos y nombres químicos
- **echa_clean** -> contiene disrupores endocrinos y nombres químicos
- **pesticidas** -> contiene el listados de las sustancias activas de pesticidas 
- **cosing_clean** -> contiene ingredientes y sustancias usadas en productos cosméticos y nombres de las etiquetas 

La primera unión -> **edlist_clean** y **echa_clean** se harán por las claves: CAS no. y EC no. 

La segunda unión -> **edlist_echa** y **pesticidas** se harán por las claves: CAS no. y EC no.

La tercera unión -> **edlist_echa_pesticidas** y **cosing_clean** se harán por las claves: CAS no. y EC no. 

**IMPORTANTE:** En adelante cuando nos refiramos al **nombre INCI** nos referimos al nombre de la etiqueta de los productos, porque es como se le conoce de manera internacional

## 1. Importamos librerias y definimos funciones

import pandas as pd
import numpy as np

**DEFINIMOS FUNCIONES**

def nombre_mas_informativo_generico(df, col1, col2, nueva_col):
    def elegir_mas_informativo(row):
        val1 = row[col1]
        val2 = row[col2]

        if pd.isna(val1):
            return val2
        if pd.isna(val2):
            return val1
        if val1 == val2:
            return val1
        return val1 if len(str(val1)) > len(str(val2)) else val2

    df[nueva_col] = df.apply(elegir_mas_informativo, axis=1)
    return df

## 2. Importamos Edlist_clean

# leemos los datos de la lista de EDlist_clean
edlist_clean= pd.read_parquet("../../data/processed/notebooks/edlist_clean.parquet")

edlist_clean.info()

# Consulatamos si hay al menos un None en string en alguna columna
edlist_clean.isin([None]).any()

# Reemplazamos None por NaN
edlist_clean = edlist_clean.replace({None: np.nan})

edlist_clean['Name and abbreviation'].nunique(dropna=True)

Hay 209 disruptores en esta lista

## 3. Importamos ECHA_clean

# leemos los datos de la lista de ECHA_clean
echa_clean= pd.read_parquet("../../data/processed/notebooks/echa_clean.parquet")

echa_clean.info()

# Consulatamos si hay al menos un None en string en alguna columna
echa_clean.isin(["None"]).any()

echa_clean['Substance name'].nunique(dropna=True)

Hay 149 disruptores en esta lista

## 4. Primera unión de dataframes

La unión que se llevará a cabo es un full outer join, es decir queremos todo de ambas tablas, porque así ganamos registros 

edlist_echa = edlist_clean.merge( echa_clean, on=['CAS Number', 'EC Number'],  how='outer', suffixes=('', '_ECHA'))

edlist_echa.info()

Como hay dos columnas que informan de la fuente original pero de las distintas tablas vamos a combinar "Fuente_original" y "Fuente_original_ECHA" 

# Combinamos para que si "Fuente_original" está vacío, se rellene con "Fuente_original_ECHA"
edlist_echa['Fuente_original'] = edlist_echa['Fuente_original'].combine_first(edlist_echa['Fuente_original_ECHA'])

# Eliminar la columna sobrante
edlist_echa.drop(columns=['Fuente_original_ECHA'], inplace=True)


print(edlist_echa['Fuente_original'].info())
edlist_echa['Fuente_original'].value_counts()

De los 295 registros hay disrupotores que aun no se han confirmado y están en evaluación:

* Edlist_2                 93
* ECHA - ED_pendiente      46

## 5. Comprobaciones del Merge

Consultamos los datos para ver si la unión se ha hecho correctamente.

Para ello comprobaresmos los campos "Name and abbreviation" proveniente de la EDlist y el "Substance name" proveniente de la ECHA, si ambos están rellenos los nombres tienen que coincidir o ser muy similares en los siguientes filtros:

1. En los registros que cumplen la condición de que 'CAS no.' y 'EC no.' ambos sean NO NULOS, 
2. En los registros que cumplen la condición de que 'CAS no.' sea NO NULO y que 'EC no.' sean  NULOS
3. En los registros que cumplen la condición de que 'CAS no.' sea NULO y que 'EC no.' sean NO NULOS
4. En los registros que cumplen la condición de que 'CAS no.' y 'EC no.' ambos sean NULOS --> no tiene que haber registro con esta condición porque hemos eliminado esta condición de ambas listas en sus respectivos ETLs

#Consulta del número de registros que cumple la condición 1
Num_consulta_1= edlist_echa[(edlist_echa['CAS Number'].notna()) & (edlist_echa['EC Number'].notna())]
Num_consulta_1[['CAS Number', 'EC Number']].count()

# Consulta del listado de algunos de los registros de la condición 1
edlist_echa[(edlist_echa['CAS Number'].notna()) & (edlist_echa['EC Number'].notna())].tail(3)

#Consulta del número de registros que cumple la condición 2
Num_consulta_2= edlist_echa[(edlist_echa['CAS Number'].notna()) & (edlist_echa['EC Number'].isna())]
Num_consulta_2[['CAS Number', 'EC Number']].count()

# Consulta de algunos de los registros de la condición 2
edlist_echa[(edlist_echa['CAS Number'].notna()) & (edlist_echa['EC Number'].isna())].tail(2)

#Consulta del número de registros que cumple la condición 3
Num_consulta_3= edlist_echa[(edlist_echa['CAS Number'].isna()) & (edlist_echa['EC Number'].notna())]
Num_consulta_3[['CAS Number', 'EC Number']].count()

# Consulta de algunos de los registros de la condición 3
edlist_echa[(edlist_echa['CAS Number'].isna()) & (edlist_echa['EC Number'].notna())].tail(3)

# Consulta de algunos de los registros de la condición 4
edlist_echa[(edlist_echa['CAS Number'].isna()) & (edlist_echa['EC Number'].isna())]

Este dataframe tiene 303 registros de disruptores endrocinos. Y las consultas que hemos hecho nos dan el total de esos 303 (245+45+13)

# Consulta tipo de nombres distintos
diferentes_names = edlist_echa[
    (edlist_echa['Name and abbreviation'].notna()) &
    (edlist_echa['Substance name'].notna()) &
    (edlist_echa['Name and abbreviation'] != edlist_echa['Substance name'])
]
diferentes_names[['CAS Number', 'EC Number', 'Name and abbreviation', 'Substance name']].count()

diferentes_names[['Name and abbreviation', 'Substance name' ]].head(10)

# Seleccionamos el nombre más informativo
nombre_mas_informativo_generico(edlist_echa, 'Name and abbreviation', 'Substance name', 'Name_Edlist_Echa')

edlist_echa [['Name and abbreviation', 'Substance name', 'Name_Edlist_Echa']].iloc[20:30]

# Elegimos las variables que queremos conservar en el dataframe final
edlist_echa =  edlist_echa[['Name_Edlist_Echa','Fuente_original','Appears on lists', 'Health effects','CAS Number', 'EC Number'  ]]

edlist_echa.info()

# Consulatamos valores únicos
edlist_echa['Name_Edlist_Echa'].nunique(dropna=True)

# Consulatamos valores únicos
edlist_echa['CAS Number'].nunique(dropna=True)

# Consulatamos valores únicos
edlist_echa['EC Number'].nunique(dropna=True)

Con esto sabemos que el listado máximo que podemos tener de disruptores es de 295.

Según los identificadores hay 247 y 227 pero estos pueden estar conbinados.

## 6. Segundo Merge

La unión se realizará con un left join por 'CAS Number'

# leemos los datos de la lista de sutanncias activas
pesticidas = pd.read_parquet("../../data/processed/notebooks/pesticidas.parquet")
pesticidas.info()


# Realizamos unión con la lista de pesticidas

edlist_echa_pesticidas = edlist_echa.merge(pesticidas, on=['CAS Number'], how='left' )
edlist_echa_pesticidas.info()

# Devuelve True si hay al menos un None en string en esa columna
edlist_echa_pesticidas.isin(["None"]).any()

edlist_echa_pesticidas [(edlist_echa_pesticidas['Name_Pesticida'].notna())]

disruptores_clean = edlist_echa_pesticidas[edlist_echa_pesticidas['Name_Pesticida'].isna()]
disruptores_clean.info()

### 6.1. Exportación del Segundo Merge para otras fuentes de búsquedas del nombre INCI.

# Elegimos las variables que queremos conservar en el dataframe final
disruptores_clean = disruptores_clean[['Name_Edlist_Echa', 'Fuente_original', 'Appears on lists', 'Health effects', 'CAS Number', 'EC Number']]

# Exportamos primera unión para otro Merge posterior
disruptores_clean.to_parquet("../../data/processed/notebooks/disruptores_clean.parquet", index=False)

##  7. Importamos COSING_clean

# leemos los datos de la lista de CosIng_clean
cosing_clean= pd.read_parquet("../../data/processed/notebooks/cosing_clean.parquet")

cosing_clean.info()

# Consulatamos si hay al menos un None en string en alguna columna
cosing_clean.isin([None]).any()

#Reemplazamos None por NaN
cosing_clean = cosing_clean.replace({None: np.nan})

cosing_clean['Name of Common Ingredients Glossary'].nunique(dropna=True)

Tenemos 541 nombres de la variable que nos da el nombre INCI

## 8. Consultas antes de segunda unión de dataframes

Vamos a consultar el número de veces que se repite el valor de CAS Number y EC Number en cada tabla antes de unir, para hacernos una idea de lo que puede ocurrir con la unión.

En esta segunda unión van a haber muchos duplicados por la condición de CAS Number y EC Number en ambas tablas.

cosing_clean['CAS Number'].value_counts().head()

cosing_clean['EC Number'].value_counts().head()

edlist_echa['CAS Number'].value_counts().head()

edlist_echa['EC Number'].value_counts().head()

Consultamos tambien cuantos registros que contengan ("EC Number" o "CAS Number") de la Edlist_Echa están en CosIng_clean y viceversa

edlist_echa[
    edlist_echa["EC Number"].isin(cosing_clean["EC Number"]) |
    edlist_echa["CAS Number"].isin(cosing_clean["CAS Number"])
].count()

cosing_clean[
    cosing_clean["CAS Number" ].isin(edlist_echa["CAS Number"]) |
    cosing_clean["EC Number"   ].isin(edlist_echa["EC Number"])
    ].count()

Consultamos nulos, pues esto será un problema a la hora de la unión, pues cuando hay nulos en la clave a unir, esos nulos se emparejan de forma erronea.

Aunque la eliminación de los registros que cumplen la condición de 'CAS Number' y 'CAS Number' sean nulos se hizo en los dataframes de origen, lo comprobamos.

edlist_echa[(edlist_echa['CAS Number'].isna() & edlist_echa['EC Number'].isna())]

cosing_clean[(cosing_clean['CAS Number'].isna() & cosing_clean['EC Number'].isna())]

Como la unión que queremos hacer no es como la anterior, es decir un full join , que daba igual si una de las dos claves tenía nulos, mientras la otra no la tuviera.

En esta ocasión no podemos eliminar los nulos de cada variable, porque nos quedaríamos con la posibilidad de unir por la otra clave.

Consultamos igualmente los nulos de estas variables en cada dataframe

edlist_echa[["EC Number", "CAS Number"]].isnull().sum()


cosing_clean[["EC Number", "CAS Number"]].isnull().sum()


## 9. Tercer Merge - Unión final

**La unión se realizará de la siguiente manera:**

1. Inner join por 'CAS Number', solo nos quedamos ocn los registros coincidentes en ambas tablas. El problema es que hay CAS Number nulos que no podemos eliminar, que haran emparejamientos erroneos.
2. Eliminación de 'CAS Number' nulos.
3. Inner join por 'EC Number' pasa lo mismo que con la unión anterior.
4. Eliminación de 'EC Number' nulos.
5. Concatenación de ambas tablas.

### 9.1. Unión por CAS Number

union_inner_cas= disruptores_clean.merge( cosing_clean, on=['CAS Number'],  how='inner', suffixes=('', '_cosing'))
union_inner_cas.info()

# Devuelve True si hay al menos un None en string en esa columna
union_inner_cas.isin([None]).any()

union_inner_cas = union_inner_cas.replace({None: np.nan})

Como esperabamos, se han creado registros con CAS Number vacios.

Solo se ha realizado la unión correcta de 69 registros.

Vemos que cuando no CAS Number es no nulo, el emparejamiento es correcto, pues EC Number y EC Number_cosing coinciden y las variables relativas a nombres de ambas tablas tambien.

union_inner_cas[(union_inner_cas['CAS Number'].notna())].head()

Sin embargo cuando es CAS Number es NULO, el emparejamiento es errático, pues EC Number y EC Number_cosing NO coinciden y las variables relativas a nombres TAMPOCO

union_inner_cas[(union_inner_cas['CAS Number'].isna())].head(3)

Procedemos a eliminar las variables donde CAS Number es nulo

union_inner_cas= union_inner_cas[~(union_inner_cas['CAS Number'].isna())]

### 9.2. Unión por EC Number

union_inner_ec= edlist_echa.merge( cosing_clean, on=['EC Number'],  how='inner', suffixes=('', '_cosing'))
union_inner_ec.info()

De la misma manera, se han creado 24.138 registros pero solo son válidos 63 EC Number

union_inner_ec[(union_inner_ec['EC Number'].isna())].head(4)

union_inner_ec= union_inner_ec[~(union_inner_ec['EC Number'].isna())]

union_inner_ec.info()

### 9.3. Unión de ambas tablas

# Concatenamos
union_completa = pd.concat([union_inner_cas,union_inner_ec], ignore_index=True)

union_completa.info()

union_completa['Name_Edlist_Echa'].nunique(dropna=True)

union_completa['Name of Common Ingredients Glossary'].nunique(dropna=True)

De los 126 registros hay 56 registros únicos, de los cuales 17 registros son de la variable que nos interesa 'Name of Common Ingredients Glossary'.

El resto son duplicados

## 10. Comprobaciones del 3º Merge

Para verificar que la unión se ha realizado correctamente, vamos a comprobar que los CAS Number y EC Number originales coindidan con los creados en la unión CAS Number_cosing y EC Number_cosing

Consulta_1 = union_completa[(union_completa['EC Number'].notna() & union_completa['EC Number_cosing'].notna())]
Consulta_1[['EC Number', 'EC Number_cosing']].head()

Consulta_2 = union_completa[(union_completa['CAS Number'].notna() & union_completa['CAS Number_cosing'].notna())]
Consulta_2[['CAS Number', 'CAS Number_cosing',]].head()

A simple vista parece que coinciden las variables en origen con las variables creadas.

Ahora vamos a comprobar si hay valores distintos y cuál es su comportamiento.

Para ello, vamos a filtrar por CAS/EC Number distinto pero vamos a ver las variables nombre, tanto las que proceden de la tabla anterior edlist_echa como de cosing. De esta forma podemos ver si tienen el mismo nombre aunque tengan distintos CAS/EC Number

# Consulta para CAS Number distintos
diferentes_cas = union_completa[
    (union_completa['CAS Number'].notna()) &
    (union_completa['CAS Number_cosing'].notna()) &
    (union_completa['CAS Number'] != union_completa['CAS Number_cosing'])
]

diferentes_cas[['CAS Number', 'CAS Number_cosing','EC Number', 'EC Number_cosing', 'Name_Edlist_Echa', 'Chemical name', 
'Chemical/IUPAC Name', 'Name of Common Ingredients Glossary', 'Identified INGREDIENTS or substances e.g.' ]].head(3)

Parece que el primer registro de la consulta no se ha realizado bien la unión. 

Las 4 variables que hacen referencia al nombre no coinciden 

Consultamos los nombres de las 2 tablas.

A simple vista parece que no, pero consultando el EC Number por el que se ha unido parece que se trata de la misma sustancia si consultamos la fuente de origen cosing_clean.

Se trata del mismo producto, que tiene mismo EC Number y distinto CAS Number, esto es debido al método explode utilizado para poder tener un solo número por fila.

cosing_clean.loc[cosing_clean["EC Number"] == "253-242-6", 
                 ['CAS Number','Chemical name']]


# Consulta para EC Number distintos
diferentes_ec = union_completa[
    (union_completa['EC Number'].notna()) &
    (union_completa['EC Number_cosing'].notna()) &
    (union_completa['EC Number'] != union_completa['EC Number_cosing'])
]

diferentes_ec[['CAS Number', 'CAS Number_cosing','EC Number', 'EC Number_cosing', 'Name_Edlist_Echa', 'Chemical name','Chemical/IUPAC Name', 'Name of Common Ingredients Glossary', 'Identified INGREDIENTS or substances e.g.' ]]

En este caso, parece todo en orden.

nombre_mas_informativo_generico(union_completa, 'Chemical name', 'Chemical/IUPAC Name', 'Name_Chemical_Cosing')
union_completa[['Chemical name', 'Chemical/IUPAC Name', 'Name_Chemical_Cosing']].iloc[17:26]

He consultado manualmente en la web de COSING y cuando no hay nombre "Name of Common Ingredients Glossary" su nombre INCI es el de la variable "Identified INGREDIENTS or substances e.g."

# Combinamos las vatiables de nombre 
union_completa["nombre_etiqueta"] = union_completa["Name of Common Ingredients Glossary"].combine_first(
    union_completa["Identified INGREDIENTS or substances e.g."]
)

# Elegimos las variables que hacen referencia a nombres y las que nos van a servir más adelante para otros fines.

union_completa =  union_completa[['Name_Edlist_Echa', 'Fuente_original', 'Appears on lists', 'Health effects','CAS Number', 'EC Number', 
'Name_Chemical_Cosing', 'nombre_etiqueta', 'Product Type, body parts','Anexo_cosIng' ]]

union_completa.info()

# Devuelve True si hay al menos un None en string en esa columna
union_completa.isin(["None"]).any()

## 11. Variable "nombre_etiqueta": tratamiento de símbolos y duplicados

Ahora vamos a comprobar la varible 'nombre_etiqueta', porque es la variable objetivo, pues es de la que obtendremos el nombre de la etiqueta.

Vamos a ver si esta variable tiene más de un nombre en un registro separado por comas, barras o punto y coma.

Los guiones y las comas no podemos tenerlos en cuenta porque claramente forman parte del nombre, y aunque parece más un nombre químico se encuentra en esta variable.

union_completa['nombre_etiqueta'].value_counts()

A simple vista vemos que hay ingredientes en **nombre_etiqueta* que contienen más de un ingrediente pero están separados por el símbolo /

Vamos a dividir y explotar la variable

# Separamos nombre_etiqueta por /
union_completa['nombre_etiqueta'] = union_completa['nombre_etiqueta'].str.split(r'[/]')

# Descomponemos fila por nombre_etiqueta
union_completa = union_completa.explode('nombre_etiqueta')

union_completa['nombre_etiqueta'].value_counts()

union_completa.isnull().sum()

# pasamos a minusculas la variable nombre_etiqueta
union_completa['nombre_etiqueta'] = union_completa['nombre_etiqueta'].str.lower()


union_completa = union_completa.drop_duplicates(subset='nombre_etiqueta', keep='first')

## 12. Lista de disruptores con nombre de etiqueta de este proceso de ETL

union_completa.info()

**LISTA DEFINITIVA DE ESTE ETL**

disruptores_etiqueta = union_completa[(union_completa['nombre_etiqueta'].notnull())]

# Dejamos las variables que necesitamos
disruptores_etiqueta = disruptores_etiqueta[['Name_Edlist_Echa', 'Name_Chemical_Cosing','Fuente_original','Anexo_cosIng','Appears on lists', 'Health effects','CAS Number','EC Number','nombre_etiqueta']]
disruptores_etiqueta.info()

# Exportamos 
disruptores_etiqueta.to_excel("../../data/processed/notebooks/disruptores_etiqueta.xlsx", index=False)
disruptores_etiqueta.to_parquet("../../data/processed/notebooks/disruptores_etiqueta.parquet", index=False)
