==> tfm-disruptors/notebooks/etl/06_Lista_definitiva.ipynb <==
# UNIÓN DE TABLAS - BUSQUEDA MANUAL
El objetivo de este archivo es continuar con la busqueda del nombre INCI que es la que aparecerá en los productos. 
Partimos de la base de que tenemos una lista de disruptores endocrinos, pero dichos disruptores pueden ser sustancias o ingredientes utilizadas para la elaboración de:
- productos cosméticos.
- pesticidas
- fungicidas
- productos de limpieza
- y un largo etc...

El cruce con la base de datos de sustancias activas de pesticidas no ha sido suficiente.

Para nuestra búsqueda de nombres INCI vamos a realizar los siguientes pasos:

1. Unión nuevamente de la tabla de disruptores limpia de pesticidas **"disruptores_clean"** con la tabla que tienen los nombres de las etiquetas **"disruptores_etiqueta"**, para poder filtrar las que faltan por rellenar.
2. Busqueda manual de nombres INCI en varias fuentes que no disponen ni de API ni de archivos para hacer merges
3. Tratamiento y exportación de listado final para tratarlo con NLP

La primera unión -> **disruptores_clean** y **disruptores_etiqueta**: 

* Se harán por las claves: CAS no. y EC no.
* Objetivo: Unir para quitar de diruptores_clean los que ya tienen etiqueta y con ello hacer búsqueda manual.


## 1. Importamos librerias

import pandas as pd
import numpy as np

## 2. Lectura de datos de "disrupores_clean" y "disruptores_etiqueta"

# leemos datasets de disrupores_clean
disruptores_clean = pd.read_parquet("../../data/processed/notebooks/disruptores_clean.parquet")
disruptores_clean.info()

# leemos datasets de disrupores_etiqueta
disruptores_etiqueta = pd.read_parquet("../../data/processed/notebooks/disruptores_etiqueta.parquet")
disruptores_etiqueta.info()

## 2. Disruptores sin nombre INCI

# consultamos duplicados
duplicados = disruptores_clean[disruptores_clean.duplicated(subset='Name_Edlist_Echa', keep=False)].sort_values(by=['CAS Number']) 
duplicados.head()

No podemos eliminar duplicados por la combinación que hay entre CAS Number y EC Number. Es mejor manterner para tener más posibilidades de encontrar el nombre de la etiqueta en otra base de datos.

# Hacemos merge entre disrupores_clean y disrupores_etiqueta
disruptores_provisional = disruptores_clean.merge(disruptores_etiqueta[[ 'CAS Number', 'EC Number','Name_Chemical_Cosing', 'nombre_etiqueta',  'Anexo_cosIng']], on=['CAS Number', 'EC Number'], how='left' )

disruptores_provisional.info()

disruptores_provisional['nombre_etiqueta'].nunique(dropna=True)

disruptores_provisional[['Name_Edlist_Echa']].nunique(dropna=True)

Como vemos la variable 'Name_Edlist_Echa' que venía sin nulos antes de la unión tiene 257 registros únicos de los cuales hay 36 registros únicos del 'nombre_etiqueta'.
Por lo que después del filtrado deberían quedar 257-36 = 221 registros sin nombre de etiqueta (eliminando duplicados obviamente)

# consultamos duplicados sin tener en cuenta el primer registro, es decir los que vamos a eliminar
duplicados = disruptores_provisional[disruptores_provisional.duplicated(subset=['Name_Edlist_Echa' ], keep='first')].sort_values(by=['nombre_etiqueta', 'Name_Chemical_Cosing']) 
duplicados.count()

298 registros totales - 257 únicos = 41 duplicados. 

disruptores_provisional = disruptores_provisional.drop_duplicates(subset=['Name_Edlist_Echa'], keep='first')

disruptores_provisional.info()

Ahora en vez de eliminar 36 registros se van a eliminar 30, por tanto obtendremos un dataframe de 257-30 = 227

disruptores_sin_etiqueta = disruptores_provisional[(disruptores_provisional['nombre_etiqueta'].isnull())]
disruptores_sin_etiqueta.info()

# Eliminamos variables que ya no tienen registros.
disruptores_sin_etiqueta = disruptores_sin_etiqueta[['CAS Number', 'EC Number', 'Name_Edlist_Echa', 'Fuente_original', 'Appears on lists','Health effects' ]]

disruptores_sin_etiqueta.head()

disruptores_sin_etiqueta['Name_Edlist_Echa'].nunique(dropna=True)

El objetivo es encontrar el nombre INCI a los 227 registros, pero sabemos que pueden haber sustancias que no se corresponden a productos cosméticos.

**EXPORTAMOS PARA SU TRATAMIENTO MANUAL**

# EXCEL
disruptores_sin_etiqueta.to_excel("../../data/processed/notebooks/disruptores_sin_etiqueta.xlsx", index=False)

## 3. Recuperación de listados con tratemiento manual con nombre INCI

Se han realizado tratamientos manuales de los dos archivos, 

# leemos datasets de disrupores_clean
inci_manual = pd.read_excel("../../data/processed/notebooks/consultas_manuales/disruptores_sin_etiqueta_manual.xlsx")

# leemos datasets de disrupores_clean
inci = pd.read_excel("../../data/processed/notebooks/consultas_manuales/disruptores_etiqueta_manual.xlsx")


## 4. Tratamiento de "inci_manual"

Este dataframe tiene más variables y más columnas de las iniciales por los siguientes motivos:
1. Se han añadido variables como:
* "nombre_etiqueta" : está es la que está en el otro dataframe en el que si se consiguió el nombre con el merge pero esta variable se ha rellenado de forma manual
* "fuente_manual", "productos_cosmeticos", "uso", "imagen_nombre", "imagen_cantidad" "texto", "repetido".
2. La variable "imagen_nombre" hace referencia al nombre de las capturas de pantalla de etiqueta de dicho disruptor, para una evaluación posterior en ocr y nlp.
3. La variable "imagen_cantidad" hace referencia a la cantidad de imágenes que se han hecho para esa sustancia
4. La variable "texto" es el texto escrito de una de esas imagenes.
5. Se han añadido filas debido a que un producto puede tomar más nombres, por lo que se ha duplicado toda la información, salvo el "nombre_etiqueta" que contiene los distintos alias. SEría como un "explode" manual.

Vamos a obtener el listado disruptores con nombre INCI, para ello:

1. Filtramos por `nombre_etiqueta`
2. Quitamos los que la variable `repetidos` ==1 identificados en la búsqueda manual porque tenían puesta la formulación distinta pero se trataba de la misma sustancia. Por ejemplo:
    * Nonylphenol, ethoxylated (7-EO)
    * 23-(nonylphenoxy)-3,6,9,12,15,18,21-heptaoxatricosan-1-ol
    * Poly (oxy-1,2-ethanediyl), alpha -(nonylphenyl)-omega-hydroxy-, branched   --> es la forma generica para el mismo y otros terminados en otros números.
3. Filtramos por la variable `productos_cosmeticos` == 1, y los que tienen valor "0" o nulo es porque o no lo son o se desconoce que se use en productos cosméticos.

inci_manual.info()

# Filtramos por nombre manual
inci_manual = inci_manual[inci_manual['nombre_etiqueta'].notnull()]
# Quitamos los repetidos
inci_manual = inci_manual[inci_manual['repetido'].isnull()]


# Filtramos 'productos_cosmeticos' que toman valor 0
consulta= inci_manual[inci_manual['productos_cosmeticos']==0]
consulta[['productos_cosmeticos', "uso"]]

inci_manual= inci_manual[inci_manual['productos_cosmeticos']==1]

inci_manual.info()

# Quitamos la variable 'repetido' porque ya no tiene sentido
inci_manual = inci_manual.drop(columns=['repetido'], errors='ignore')

Vamos a pasar las variables float a numérico pero antes vamos a convertir sus nulos en '0'

cols = ["Health_effects","productos_cosmeticos", "imagen_cantidad"]

inci_manual[cols] = inci_manual[cols].fillna(0).astype(int)

Ahora vamos a consultar en cuantos de los 227 registros de salida hemos coseguido su nombre INCI.

Aunque hayamos encontrado más de uno para alguna sustancia, lo que queremos saber de esos 227 cuales se han rellenado.

inci_manual['nombre_etiqueta'].nunique(dropna=True)

inci_manual['Name_Edlist_Echa'].nunique(dropna=True)

Esto quiere decir que de los 227 hemos conseguido rellenar 47 y que esas 47 sustancias tienen por termino medio 2 nombres, porque tenemos hasta 104 nombres de etiqueta

## 5. Tratamiento de "inci" - dataframe inicial: disruptores_etiqueta

Este dataframe tiene más variables y más columnas de las iniciales por los siguientes motivos:
1. Se han añadido variables como: 
* "uso","productos_cosmeticos", , "imagen_nombre", "imagen_cantidad", "texto".

2. Al igual que con el otro dataframe se han añadido filas debido a que un producto puede tomar más nombres.

# Consultamos la información de inci.
inci.info()

Vamos a pasar la variable float a numérico pero antes vamos a convertir sus nulos en '0'

inci["Health_effects"] = inci["Health_effects"].fillna(0).astype(int)
inci["Health_effects"] = (inci["Health_effects"] == 1).astype(int)

## 6. LISTA DEFINITIVA DE DISRUPTORES EN PRODUCTOS COSMÉTICOS

# Concatenamos los DataFrames
disruptores_final = pd.concat([inci, inci_manual], ignore_index=True)
disruptores_final.info()


# es importante que estas columnas sean de tipo string
cols = ['CAS Number', 'EC Number']
disruptores_final[cols] = (
    disruptores_final[cols]
      .apply(lambda col: col.astype('string').str.strip())
)

# las variables EC Number y Anexo_cosIng tienen valores nulos que vamos a rellenar
disruptores_final["EC Number"] = disruptores_final["EC Number"].fillna("")
disruptores_final["Anexo_cosIng"] = disruptores_final["Anexo_cosIng"].fillna("sin info")

Vamos a comprobar si el número de imagenes físicas que tenemos coincide con la suma de la unión de ambos dataframes: 
Si coincide

total_imagenes = disruptores_final["imagen_cantidad"].sum()
print(total_imagenes)

# Filtramos filas con varios nombres que no se dividió a mano, porque no había foto
disruptores_final[disruptores_final['nombre_etiqueta'].str.contains(r'[/;]', regex= True,na=False)]

# Separamos por / 
disruptores_final['nombre_etiqueta']= disruptores_final['nombre_etiqueta'].str.split(r'[/]')

# Descomponemos fila por cada nombre
disruptores_final = disruptores_final.explode('nombre_etiqueta')

disruptores_final.iloc[69:75, 6:9]

disruptores_final.iloc[108:112, 6:9]

Para análisis posteriores es importante que las variables Fuente_original y Anexo_cosing muestren más información:

disruptores_final["Fuente_original"].value_counts()

# Diccionario de mapeo
map_fuente = {
    "Edlist_1": "EDlist_1 (confirmado UE)",
    "Edlist_2": "EDlist_2 (evaluando)",
    "Edlist_3": "EDlist_3 (confirmado país miembro, pero no UE)",
    "ECHA - ED_pendiente": "ECHA (evaluando o pendiente)",
    "ECHA - ED_Env": "ECHA (efectos ambiente)",
    "ECHA - ED_Humans&Env": "ECHA (efectos salud y ambiente)"
}

# Aplicar el mapeo en la columna Fuente_original
disruptores_final["Fuente_original"] = disruptores_final["Fuente_original"].replace(map_fuente)

disruptores_final["Anexo_cosIng"].value_counts()

map_cosing = {
    "Anexo_2": "Anexo II (Ingrediente Prohibido)",
    "Anexo_3": "Anexo III (Ingrediente Restringido)",
    "Anexo_4": "Anexo IV (Colorante Permitido)",
    "Anexo_5": "Anexo V (Conservante Permitido)",
    "Anexo_6": "Anexo VI (Filtro UV Permitido)",
    "sin info": "Sin información"
}

disruptores_final["Anexo_cosIng"] = disruptores_final["Anexo_cosIng"].replace(map_cosing)

disruptores_final.info()

n_dis_inicial=disruptores_provisional.shape[0]
n_alias = disruptores_final["nombre_etiqueta"].nunique()
n_cas = disruptores_final["CAS Number"].nunique()
n_ec = disruptores_final["EC Number"].nunique()
n_name_edlist = disruptores_final["Name_Edlist_Echa"].nunique()
n_texto = disruptores_final["texto"].nunique()

print(f"Número de Disruptores iniciales antes de incluir nombre de etiqueta: {n_dis_inicial}")
print(f"Alias totales conseguidos (filas de nombre_etiqueta): {n_alias}")
print(f"CAS totales (registros CAS Number único): {n_cas}")
print(f"EC totales (registros EC Number único): {n_ec}")
print(f"Disruptores con nombre INCI (registros Name_Edlist_Echa único): {n_name_edlist}")
print(f"Textos totales (registros texto único): {n_texto}")

Hay menos **Name_Edlist_Echa** que **CAS y EC** porque hay ingredientes que tiene más de un CAS y EC por ejemplo:

disruptores_final[disruptores_final["Name_Edlist_Echa"] == "Cyclomethicone"]

### 6.1. Conclusión
**Hemos obtenido un total de 77 de 257 disruptores con su nombre INCI, es decir un 30% del total de disruptores, el resto, pueden ser sustancias usadas en productos de limpieza, pesticidas, herbicidas, etc.., como se ha podido observar en la importación de la busqueda manual, y de los restantes no sabemos qué tipo de sustancias son.**

## 7. Exportamos la lista de diruptores final
Se exporta lista para usarla en el nlp. Tenemos en total 161 nombres de etiquetas

# Reiniciamos ínfice
disruptores_final.reset_index(drop=True, inplace=True)

disruptores_final.to_parquet("../../data/processed/notebooks/disruptores_final.parquet", index=False)

disruptores_final.to_excel("../../data/processed/notebooks/disruptores_final.xlsx", index=False)
